ETH-T17: strict BigDecimal financial engine

+++ java-service/src/main/java/com/ethsimulator/agent/dto/ParseGoalResponse.java
@@ -1,5 +1,6 @@
 package com.ethsimulator.agent.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -7,7 +8,7 @@ public record ParseGoalResponse(
         String intent,
         Map<String, Object> entities,
         String suggestedAction,
-        double confidence,
+        BigDecimal confidence,
         List<Map<String, Object>> toolsUsed,
         List<Map<String, Object>> chartSpecs,
         List<Map<String, Object>> feedback
