+++ java-service/src/main/java/com/ethsimulator/wallet/StablecoinBalance.java
@@ -1,10 +1,12 @@
 package com.ethsimulator.wallet;
 
+import java.math.BigDecimal;
+
 public record StablecoinBalance(
         String symbol,
         String contractAddress,
         int decimals,
         String balance,
-        double balanceUsd
+        BigDecimal balanceUsd
 ) {
 }
